pyproject.toml
[project]
name = "whisper-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.77.0",
    "pyannote-audio>=3.3.2",
    "python-dotenv>=1.1.0",
]
